refactor: the cli is now in iroh one
