+++ iroh-one/src/lib.rs
@@ -8,6 +8,7 @@ pub mod gateway_cli; // XXX remove pub once move is complete
 pub mod mem_p2p;
 pub mod mem_store;
 pub mod p2p_cli;
+mod start;
 pub mod status; // XXX remove pub once move is complete
 pub mod store_cli; // XXX remove pub once move is complete
 #[cfg(feature = "uds-gateway")]
